improved ExecuteCommandProvider and added unit test for EventLoop

+++ root/src/main/java/com/github/pedroarrudamoreira/vaultage/root/backup/provider/impl/ExecuteCommandProvider.java
@@ -4,12 +4,14 @@ import java.io.File;
 import java.io.InputStream;
 import java.io.OutputStream;
 import java.util.List;
+import java.util.concurrent.TimeUnit;
 
 import com.github.pedroarrudamoreira.vaultage.process.ProcessSpawner;
 import com.github.pedroarrudamoreira.vaultage.root.backup.provider.BackupProvider;
 import com.github.pedroarrudamoreira.vaultage.root.backup.provider.util.BackupProviderUtils;
 import com.github.pedroarrudamoreira.vaultage.root.security.model.User;
 import com.github.pedroarrudamoreira.vaultage.root.util.RootObjectFactory;
+import com.github.pedroarrudamoreira.vaultage.util.EventLoop;
 import com.github.pedroarrudamoreira.vaultage.util.IOUtils;
 import com.github.pedroarrudamoreira.vaultage.util.ObjectFactory;
 
@@ -23,11 +25,10 @@ public class ExecuteCommandProvider implements BackupProvider {
 	public void doBackup(User user, InputStream database, Object params) {
 		File tempFolder = null;
 		File tempFile = null;
-		File theFile = null;
 		try {
 			tempFile = RootObjectFactory.buildTempFile("backup", ".zip");
 			tempFolder = tempFile.getParentFile();
-			theFile = ObjectFactory.buildFile(tempFolder, BackupProviderUtils.createFileName(user));
+			final File theFile = ObjectFactory.buildFile(tempFolder, BackupProviderUtils.createFileName(user));
 			@Cleanup OutputStream out = ObjectFactory.buildFileOutputStream(theFile);
 			IOUtils.copy(database, out);
 			out.close();
@@ -35,13 +36,18 @@ public class ExecuteCommandProvider implements BackupProvider {
 			for(int i = 0; i < paramsList.size(); i++) {
 				paramsList.set(i, String.format(paramsList.get(i), theFile.getAbsolutePath()));
 			}
-			ProcessSpawner.executeProcessAndWait(paramsList.toArray(new String[paramsList.size()]));
+			final Process process = ProcessSpawner.executeProcess(null, paramsList.toArray(new String[paramsList.size()]));
+			EventLoop.repeatTask(() -> {
+				if(!process.isAlive()) {
+					if(theFile != null) {
+						theFile.delete();
+					}
+				}
+				return process.isAlive();
+			}, 2, TimeUnit.SECONDS);
 		} finally {
 			if(tempFile != null) {
 				tempFile.delete();
-			}
-			if(theFile != null) {
-				theFile.delete();
 			}
 		}
 	}
